Change format of metadata mapping file to emit correct RDF
The earlier template for statements omitted the Q ID part that was prepended to the statement UUIDs. This was added in the examples.

Also, the template for full IRI values was adding a base IRI and URL-encoding the actual IRI. This was prevented by adding a + in front of the variable name in the curly brackets.